+++ CHANGELOG.md
@@ -3,6 +3,8 @@ LUYA CHANGELOG
 
 1.0.0-beta3 (in progress)
 -------------------------
+
+- `#634` added getContainer and getLevelCurrentItems for menu component.
 - `#633` added the ability to register new translations/message quickly in the Module.php via $translation property.
 
 1.0.0-beta2 (3.12.2015)